added sweetalert for a successfull delete oparation

+++ src/components/main/MyCampaign.jsx
@@ -1,15 +1,41 @@
+import { useState } from "react";
 import { Link, useLoaderData } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const MyCampaign = () => {
-  const data = useLoaderData();
+  const loadedData = useLoaderData();
+  const [data, setData] = useState(loadedData);
+  // setData(fetchData);
   // console.log(data);
+
   const handleDelete = (id) => {
-    console.log(id);
-    fetch(`http://localhost:5000/campaign/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => console.log(data));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/campaign/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((result) => {
+            if (result.deletedCount) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+              const remainCampaign = data.filter((data) => data._id !== id);
+              setData(remainCampaign);
+            }
+          });
+      }
+    });
   };
 
   return (
